refactor: remove GetAllTxns, rename next txns tests
